requirements.txt
Flask
gunicorn
sqlalchemy
flask-sqlalchemy
psycopg2-binary
python-dotenv
flask-admin
flask-wtf
flask-migrate
flask-babel
flask-babelEx